poc/pyproject.toml
[project]
name = "ai-extract-poc"
version = "1.0.0"
description = "AI_EXTRACT POC Kit — Tests"
requires-python = ">=3.11"
dependencies = [
    "streamlit[snowflake]==1.50.0",
    "plotly>=5.18.0",
    "pypdfium2>=4.0",
    "snowflake-snowpark-python>=1.20.0",
]

[dependency-groups]
dev = [
    "playwright>=1.58.0",
    "pytest>=9.0.1",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.5.0",
    "snowflake-connector-python>=3.6.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-ra", "-q"]
markers = [
    "smoke: core happy-path tests that should always pass quickly",
    "slow: tests that take >10s (e.g. full navigation, multi-page)",
    "sql: SQL integration tests against live Snowflake",
    "e2e: end-to-end Streamlit browser tests",
]

[tool.coverage.run]
source = ["streamlit"]
omit = ["tests/*", "conftest.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
